# HttpExample/18.HttpHealthcheck/obj/Debug/net8.0/EndpointInfo/7.Swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Todo API",
    "description": "간단한 Todo 관리용 Web API",
    "contact": {
      "name": "김ss",
      "email": "you@example.com"
    },
    "version": "v1"
  },
  "paths": {
    "/api/v1/Todo": {
      "get": {
        "tags": [
          "Todo"
        ],
        "summary": "전체 Todo 목록 조회",
        "description": "DB에 저장된 모든 Todo 항목을 조회합니다.",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetTodoDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetTodoDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetTodoDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Todo"
        ],
        "summary": "새로운 Todo 추가",
        "description": "Title을 입력 받아 새로운 Todo를 생성합니다.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성 성공",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoDto"
                }
              }
            }
          },
          "400": {
            "description": "입력 값 검증 실패"
          }
        }
      }
    },
    "/api/v1/Todo/{id}": {
      "get": {
        "tags": [
          "Todo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Todo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Todo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateTodoDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetTodoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UpdateTodoDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}